+++ js/string-length.js
@@ -1,0 +1,7 @@
+// Проверка максимальной длины строки
+
+function validateStringLength (string, maxLength) {
+  return string.length <= maxLength;
+}
+
+export {validateStringLength};
