+++ js/until.js
@@ -1,0 +1,13 @@
+const getRandomInt = (a, b) => {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (array) => {
+  return array[getRandomInt(0, array.length - 1)]
+}
+
+export {getRandomArrayElement};
+export {getRandomInt};
